Display notifications on success
Readd notify-helper
